added valid_session_table as a property of sessionfilter
If get_method consults bosf.bocCache.tAllsessions, failed sessions will be included. If get_method consults filtered_session_table, the result will vary. In this sense, it is necessary to have a "valid_session_table". It can be hidden, but I personally want to see it there. Tutorials are updated accordingly.