modified the livescript as Vijay suggested
I didn't change the original filtered session but simply added this specific session for comparison. I think it's also important for users to know that not all mice ran. Besides, if it comes up that this specific session is assessed "failed" sometime later, the livescript can still run (keep things hard coded as few as possible).