Update Unit.m
Explicitly create a linked session for unit as an ephys session. If the ophys manifest is missing from the cache, calling bot.session will trigger fetching of the ophys manifest in order to check if the provided session id is an ophys session. That is not necessary here, because a unit is always derived from an ephys session.